dashboard/requirements.txt
attrdict
flask
gevent
termcolor
pyfiglet
psycopg2
pyyaml